+++ order/src/main/java/com/watermelon/server/order/TokenProviderConfig.java
@@ -1,0 +1,9 @@
+package com.watermelon.server.order;
+
+
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class TokenProviderConfig {
+
+}
