+++ order/src/main/java/com/watermelon/server/order/service/orderApplyCount/OrderApplyCountLockService.java
@@ -1,6 +1,6 @@
-package com.watermelon.server.orderApplyCount.service;
+package com.watermelon.server.order.service.orderApplyCount;
 
-import com.watermelon.server.orderApplyCount.domain.OrderApplyCount;
+import com.watermelon.server.order.domain.OrderApplyCount;
 import org.springframework.stereotype.Service;
 
 @Service
