--- order/src/main/java/com/watermelon/server/orderResult/config/TokenProviderConfig.java
@@ -1,11 +1,0 @@
-package com.watermelon.server.orderResult.config;
-
-
-import com.watermelon.server.orderResult.service.BlockingQueueTokenForDbAccessProviderService;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class TokenProviderConfig {
-
-}
