+++ order/src/test/java/com/watermelon/server/order/controller/OrderEventControllerTest.java
@@ -14,7 +14,7 @@ import com.watermelon.server.order.exception.NotDuringEventPeriodException;
 import com.watermelon.server.order.exception.WinnerAlreadyParticipateException;
 import com.watermelon.server.order.exception.WrongOrderEventFormatException;
 import com.watermelon.server.order.exception.WrongPhoneNumberFormatException;
-import com.watermelon.server.orderResult.service.OrderApplyService;
+import com.watermelon.server.order.service.orderApply.OrderApplyService;
 import com.watermelon.server.order.service.OrderEventCommandService;
 import com.watermelon.server.order.service.OrderEventQueryService;
 import org.junit.jupiter.api.BeforeEach;
