+++ order/src/test/java/com/watermelon/server/order/result/service/OrderApplyResultCommandServiceTest.java
@@ -4,10 +4,10 @@ import com.watermelon.server.order.domain.ApplyTicketStatus;
 import com.watermelon.server.order.dto.request.RequestAnswerDto;
 import com.watermelon.server.order.exception.NotDuringEventPeriodException;
 import com.watermelon.server.order.exception.WrongOrderEventFormatException;
-import com.watermelon.server.orderResult.repository.OrderApplyResultRepository;
-import com.watermelon.server.orderResult.service.MemoryOrderEventService;
-import com.watermelon.server.order.service.OrderResultSaveService;
-import com.watermelon.server.orderResult.service.OrderApplyService;
+import com.watermelon.server.order.repository.OrderApplyResultRepository;
+import com.watermelon.server.order.service.MemoryOrderEventService;
+import com.watermelon.server.order.service.orderApply.OrderResultSaveService;
+import com.watermelon.server.order.service.orderApply.OrderApplyService;
 import com.watermelon.server.token.ApplyTokenProvider;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
