+++ order/src/test/java/com/watermelon/server/order/service/MemoryOrderEventServiceTest.java
@@ -4,11 +4,10 @@ import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.dto.request.RequestOrderEventDto;
 import com.watermelon.server.order.dto.request.RequestOrderRewardDto;
 import com.watermelon.server.order.dto.request.RequestQuizDto;
-import com.watermelon.server.orderApplyCount.repository.OrderApplyCountRepository;
+import com.watermelon.server.order.repository.OrderApplyCountRepository;
 import com.watermelon.server.order.repository.OrderEventRepository;
-import com.watermelon.server.orderApplyCount.domain.OrderApplyCount;
-import com.watermelon.server.orderApplyCount.service.OrderApplyCountService;
-import com.watermelon.server.orderResult.service.MemoryOrderEventService;
+import com.watermelon.server.order.domain.OrderApplyCount;
+import com.watermelon.server.order.service.orderApplyCount.OrderApplyCountService;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
