+++ order/src/test/java/com/watermelon/server/order/service/OrderEventCommandServiceTest.java
@@ -3,8 +3,7 @@ package com.watermelon.server.order.service;
 import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.domain.Quiz;
 import com.watermelon.server.order.repository.OrderEventRepository;
-import com.watermelon.server.orderResult.service.MemoryOrderEventService;
-import com.watermelon.server.orderResult.service.OrderApplyService;
+import com.watermelon.server.order.service.orderApply.OrderApplyService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.extension.ExtendWith;
