+++ order/src/test/java/com/watermelon/server/order/total/OrderEventTotalTest.java
@@ -7,10 +7,10 @@ import com.watermelon.server.order.domain.OrderEvent;
 import com.watermelon.server.order.domain.OrderEventStatus;
 import com.watermelon.server.order.domain.Quiz;
 import com.watermelon.server.order.dto.request.*;
-import com.watermelon.server.orderApplyCount.repository.OrderApplyCountRepository;
+import com.watermelon.server.order.repository.OrderApplyCountRepository;
 import com.watermelon.server.order.repository.OrderEventRepository;
-import com.watermelon.server.orderApplyCount.domain.OrderApplyCount;
-import com.watermelon.server.orderResult.service.MemoryOrderEventService;
+import com.watermelon.server.order.domain.OrderApplyCount;
+import com.watermelon.server.order.service.MemoryOrderEventService;
 import com.watermelon.server.token.ApplyTokenProvider;
 import com.watermelon.server.token.JwtPayload;
 import lombok.extern.slf4j.Slf4j;
